// src/api/generated/model/metricLayoutMetricType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Cosy API
 * Management API for Cosy (Cost Optimised Server Yard).
 * OpenAPI spec version: v1.0
 */

export type MetricLayoutMetricType = typeof MetricLayoutMetricType[keyof typeof MetricLayoutMetricType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MetricLayoutMetricType = {
  CPU_PERCENT: 'CPU_PERCENT',
  MEMORY_LIMIT: 'MEMORY_LIMIT',
  MEMORY_PERCENT: 'MEMORY_PERCENT',
  MEMORY_USAGE: 'MEMORY_USAGE',
  NETWORK_INPUT: 'NETWORK_INPUT',
  NETWORK_OUTPUT: 'NETWORK_OUTPUT',
  BLOCK_READ: 'BLOCK_READ',
  BLOCK_WRITE: 'BLOCK_WRITE',
} as const;